# [ENH] Speed up spatial normalization (take 2)