--- a/mriqc/interfaces/__init__.py
+++ b/mriqc/interfaces/__init__.py
@@ -12,4 +12,4 @@ from mriqc.interfaces.anatomical import StructuralQC, ArtifactMask, ComputeQI2
 from mriqc.interfaces.functional import FunctionalQC, Spikes
 from mriqc.interfaces.bids import ReadSidecarJSON, IQMFileSink
 from mriqc.interfaces.viz import PlotMosaic, PlotContours, PlotSpikes
-from mriqc.interfaces.common import ConformImage
+from mriqc.interfaces.common import ConformImage, EnsureSize