… mriqc/data/testdata/T1w.csv
… setup.py
… tests/nii_T1w.txt
… tests/nii_bold.txt