@oesteban 2017-05-12
Finally passing! The performance with --ants-nthreads=1 is about the same as for 4.